GRAPHICS: Fix negation of unsigned int warnings

+++ graphics/blit/blit-alpha.cpp
@@ -151,9 +151,9 @@ static inline bool alphaBlitHelper(byte *dst, const byte *src, const byte *mask,
 	if (flipy && flipx)
 		dstDelta = -dstDelta;
 	else if (flipy)
-		dstDelta = -((dstPitch * 2) - dstDelta);
+		dstDelta = -(static_cast<int>(dstPitch * 2) - dstDelta);
 	else if (flipx)
-		dstDelta =   (dstPitch * 2) - dstDelta;
+		dstDelta = static_cast<int>(dstPitch * 2) - dstDelta;
 
 	if (aMod == 0)
 		return true;
@@ -209,9 +209,9 @@ static inline bool alphaBlitMapHelper(byte *dst, const byte *src, const byte *ma
 	if (flipy && flipx)
 		dstDelta = -dstDelta;
 	else if (flipy)
-		dstDelta = -((dstPitch * 2) - dstDelta);
+		dstDelta = -(static_cast<int>(dstPitch * 2) - dstDelta);
 	else if (flipx)
-		dstDelta =   (dstPitch * 2) - dstDelta;
+		dstDelta = static_cast<int>(dstPitch * 2) - dstDelta;
 
 	// TODO: optimized cases for dstDelta of 0
 	if (dstFmt.bytesPerPixel == 2) {
